refactor: 크롤링 데이터 전처리 리팩터링
